feat: Make code editor clickable to request focus
The code editor is now clickable, allowing users to request focus directly on the editor area. This is achieved by adding a `clickable` modifier to the `Box` containing the code editor. A `FocusRequester` is also introduced to programmatically request focus when the editor is clicked. This improves usability by making it easier to start typing in the code editor.